Moves date util functions to core
