+++ src/components/date-display/date-display.test.tsx
@@ -1,8 +1,9 @@
-import { DateDisplay, formatDateDefault } from "./date-display";
 import { describe, it } from "vitest";
 import { render, screen } from "~/test-utils";
 
+import { DateDisplay } from "./date-display";
 import React from "react";
+import { formatDate } from "~/core/date";
 
 describe("<DateDisplay />", () => {
   describe("defaults", () => {
@@ -67,7 +68,7 @@ describe("<DateDisplay />", () => {
         <DateDisplay
           className="test-date"
           value="2022-01-01"
-          formatDate={(d) => formatDateDefault(d, { year: "2-digit" })}
+          formatDate={(d) => formatDate(d, { year: "2-digit" })}
         />
       );
       expect(
