+++ src/core/date/index.ts
@@ -1,0 +1,1 @@
+export * from "~/core/date/date-core";
